Backend/myenv/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.114.0
h11==0.14.0
idna==3.8
jsonify==0.5
motor==3.5.1
pydantic==2.9.0
pydantic-core==2.23.2
PyJWT==2.8.0
pymongo==4.8.0
python-decouple==3.8
sniffio==1.3.1
starlette==0.38.5
typing-extensions==4.12.2
uvicorn==0.30.6
